Exceptions on Get, put and delete

+++ src/main/java/com/sipriano/desafio_tres/services/ClientService.java
@@ -3,6 +3,8 @@ package com.sipriano.desafio_tres.services;
 import com.sipriano.desafio_tres.entities.Client;
 import com.sipriano.desafio_tres.entities.ClientDTO;
 import com.sipriano.desafio_tres.repository.ClientRepository;
+import com.sipriano.desafio_tres.services.exceptions.IdNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -23,13 +25,13 @@ public class ClientService {
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
         Optional<Client> clientOptional = repository.findById(id);
-        Client client = clientOptional.orElseThrow(() -> new NoSuchElementException("Id não encontrado!"));
+        Client client = clientOptional.orElseThrow(() -> new IdNotFoundException("Id não encontrado!"));
         return new ClientDTO(client);
     }
 
     @Transactional(readOnly = true)
     public Page<ClientDTO> findAll(Pageable pageable) {
-         return repository.findAll(pageable).map(ClientDTO::new);
+        return repository.findAll(pageable).map(ClientDTO::new);
     }
 
     @Transactional
@@ -41,13 +43,20 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO clientDTO) {
-        Client client = repository.getReferenceById(id);
-        copyDTOToClient(clientDTO, client);
-        return new ClientDTO(repository.save(client));
+        try {
+            Client client = repository.getReferenceById(id);
+            copyDTOToClient(clientDTO, client);
+            return new ClientDTO(repository.save(client));
+        } catch (EntityNotFoundException e) {
+            throw new IdNotFoundException("Id não encontrado!");
+        }
     }
 
     @Transactional
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new IdNotFoundException("Id não encontrado!");
+        }
         repository.deleteById(id);
     }
